+++ ios/cupertino_native_extra/Sources/cupertino_native_extra/Factories/CupertinoNavigationBarFactory.swift
@@ -3,9 +3,11 @@ import UIKit
 
 public class CupertinoNavigationBarViewFactory: NSObject, FlutterPlatformViewFactory {
   private let messenger: FlutterBinaryMessenger
+  private let registrar: FlutterPluginRegistrar
 
-  init(messenger: FlutterBinaryMessenger) {
+  init(messenger: FlutterBinaryMessenger, registrar: FlutterPluginRegistrar) {
     self.messenger = messenger
+    self.registrar = registrar
     super.init()
   }
 
@@ -14,6 +16,6 @@ public class CupertinoNavigationBarViewFactory: NSObject, FlutterPlatformViewFac
   }
 
   public func create(withFrame frame: CGRect, viewIdentifier viewId: Int64, arguments args: Any?) -> FlutterPlatformView {
-    return CupertinoNavigationBarPlatformView(frame: frame, viewId: viewId, args: args, messenger: messenger)
+    return CupertinoNavigationBarPlatformView(frame: frame, viewId: viewId, args: args, messenger: messenger, registrar: registrar)
   }
 }