+++ macos/Classes/Factories/CupertinoNavigationBarFactory.swift
@@ -3,9 +3,11 @@ import Cocoa
 
 public class CupertinoNavigationBarViewFactory: NSObject, FlutterPlatformViewFactory {
   private let messenger: FlutterBinaryMessenger
+  private let registrar: FlutterPluginRegistrar
 
-  init(messenger: FlutterBinaryMessenger) {
+  init(messenger: FlutterBinaryMessenger, registrar: FlutterPluginRegistrar) {
     self.messenger = messenger
+    self.registrar = registrar
     super.init()
   }
 
@@ -14,6 +16,6 @@ public class CupertinoNavigationBarViewFactory: NSObject, FlutterPlatformViewFac
   }
 
   public func create(withViewIdentifier viewId: Int64, arguments args: Any?) -> NSView {
-    return CupertinoNavigationBarNSView(viewId: viewId, args: args, messenger: messenger)
+    return CupertinoNavigationBarNSView(viewId: viewId, args: args, messenger: messenger, registrar: registrar)
   }
 }